DRTII-894 Refactor to prod code
